perf: produce tasks lazily and consume results as made available
This change dramatically reduces the total size of objects on the heap
at any one time for webpack configurations with many entries. Because
`tasks` used to be computed eagerly and `completedTasks` was collected
as a large Array, space complexity used to be linear with respect to the
number of entries. (More precisely, the max heap size was proportional
to the combined size in bytes of all pre-minified sources).

Now, we defer the generation of a `task` (and thus the memory allocation
required for computing `asset.source()` of an entry) until a worker is
made available. Similarly, the computation of `serialize(task)`, another
large String, is deferred until a worker is available. Finally, when a
`task` is completed, the `completedTask` is consumed immediately,
releasing the reference to the original `asset.source()`, and making it
a candidate for garbage collection.

The effect is that space complexity is now roughly linear with respect
to the number of parallel workers.